update: delegate channel filtering to IFilterChannels
Took 22 minutes
