+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channels/loadmessages/SearchMessagesForQuery.kt
@@ -16,14 +16,13 @@ internal class SearchMessagesForQuery(
     private val iHelpSearchWithDebounce: IHelpSearchWithDebounce,
 ) : IChannelViewState by channelState, IHelpSearchWithDebounce by iHelpSearchWithDebounce {
 
-    private val iLoadMessages =
-        StreamMessagesLoader(
-            iChannelViewState = channelState,
-            searchDebouncer = searchDebouncer,
-            logger = logger,
-            chatClient = chatClient,
-            channelLimit = channelLimit
-        )
+    private val iLoadMessages = StreamMessagesLoader(
+        iChannelViewState = channelState,
+        searchDebouncer = searchDebouncer,
+        logger = logger,
+        chatClient = chatClient,
+        channelLimit = channelLimit
+    )
 
     internal suspend operator fun invoke(
         query: String,
